# src/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Outpoint": {
				"type": "string"
			},
			"File": {
				"properties": {
					"hash": {
						"type": "string"
					},
					"size": {
						"type": "number",
						"format": "double"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Inscription": {
				"properties": {
					"json": {},
					"text": {
						"type": "string"
					},
					"words": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"file": {
						"$ref": "#/components/schemas/File"
					},
					"parent": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Sigma": {
				"properties": {
					"algorithm": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"signature": {
						"type": "string"
					},
					"vin": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Bsv20Status": {
				"enum": [
					-1,
					0,
					1
				],
				"type": "number"
			},
			"TxoData": {
				"properties": {
					"types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"insc": {
						"$ref": "#/components/schemas/Inscription"
					},
					"map": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"b": {
						"$ref": "#/components/schemas/File"
					},
					"sigma": {
						"items": {
							"$ref": "#/components/schemas/Sigma"
						},
						"type": "array"
					},
					"list": {
						"properties": {
							"sale": {
								"type": "boolean"
							},
							"payout": {
								"type": "string"
							},
							"price": {
								"type": "number",
								"format": "double"
							}
						},
						"type": "object"
					},
					"bsv20": {
						"properties": {
							"implied": {
								"type": "boolean"
							},
							"status": {
								"$ref": "#/components/schemas/Bsv20Status"
							},
							"amt": {
								"type": "string"
							},
							"sym": {
								"type": "string"
							},
							"tick": {
								"type": "string"
							},
							"op": {
								"type": "string"
							},
							"p": {
								"type": "string"
							},
							"id": {
								"$ref": "#/components/schemas/Outpoint"
							}
						},
						"type": "object"
					},
					"lock": {
						"properties": {
							"until": {
								"type": "number",
								"format": "double"
							},
							"address": {
								"type": "string"
							}
						},
						"required": [
							"until",
							"address"
						],
						"type": "object"
					},
					"sigil": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"opns": {
						"properties": {
							"status": {
								"type": "number",
								"format": "double"
							},
							"domain": {
								"type": "string"
							},
							"genesis": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"opnsMine": {
						"properties": {
							"pow": {
								"type": "string"
							},
							"status": {
								"type": "number",
								"format": "double"
							},
							"domain": {
								"type": "string"
							},
							"genesis": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"Origin": {
				"properties": {
					"outpoint": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"data": {
						"$ref": "#/components/schemas/TxoData"
					},
					"num": {
						"type": "number",
						"format": "double"
					},
					"map": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					"inum": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"outpoint"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Txo": {
				"properties": {
					"txid": {
						"type": "string",
						"default": ""
					},
					"vout": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"outpoint": {
						"$ref": "#/components/schemas/Outpoint"
					},
					"satoshis": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"accSats": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"owner": {
						"type": "string"
					},
					"script": {
						"type": "string"
					},
					"spend": {
						"type": "string"
					},
					"spend_height": {
						"type": "number",
						"format": "double"
					},
					"spend_idx": {
						"type": "number",
						"format": "double"
					},
					"origin": {
						"$ref": "#/components/schemas/Origin"
					},
					"height": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"idx": {
						"type": "number",
						"format": "double",
						"default": 0
					},
					"data": {
						"$ref": "#/components/schemas/TxoData"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"SortDirection": {
				"enum": [
					"asc",
					"desc",
					"ASC",
					"DESC"
				],
				"type": "string"
			},
...
			"TokenBalanceResponse": {
				"properties": {
					"listed": {
						"properties": {
							"pending": {
								"type": "string"
							},
							"confirmed": {
								"type": "string"
							}
						},
						"required": [
							"pending",
							"confirmed"
						],
						"type": "object"
					},
					"all": {
						"properties": {
							"pending": {
								"type": "string"
							},
							"confirmed": {
								"type": "string"
							}
						},
						"required": [
							"pending",
							"confirmed"
						],
						"type": "object"
					},
					"icon": {
						"type": "string"
					},
					"dec": {
						"type": "number",
						"format": "double"
					},
					"sym": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"tick": {
						"type": "string"
					}
				},
				"required": [
					"listed",
					"all"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "1sat-server-js",
		"contact": {}
	},
	"paths": {
		"/api/txos/address/{address}/unspent": {
			"get": {
				"operationId": "GetUnspentByAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "bsv20",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "origins",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "refresh",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "PostUnspentByAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "bsv20",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "origins",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "refresh",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TxoData"
							}
						}
					}
				}
			}
		},
		"/api/txos/address/{address}/history": {
			"get": {
				"operationId": "GetHistoryByAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "bsv20",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "origins",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "PostHistoryByAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "bsv20",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					},
					{
						"in": "query",
						"name": "origins",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TxoData"
							}
						}
					}
				}
			}
		},
		"/api/txos/address/{address}/balance": {
			"get": {
				"operationId": "GetBalanceByAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"format": "double"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "refresh",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/api/txos/search": {
			"get": {
				"operationId": "GetTxoSearchAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortDirection"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostTxoSearchAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortDirection"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TxoData"
							}
						}
					}
				}
			}
		},
		"/api/txos/search/unspent": {
			"get": {
				"operationId": "GetTxoSearchUnspent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortDirection"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "PostTxoSearchUnspent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tag",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 100,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "offset",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "dir",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/SortDirection"
						}
					},
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": false,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TxoData"
							}
						}
					}
				}
			}
		},
		"/api/txos/{outpoint}": {
			"get": {
				"operationId": "GetTxoByOutpoint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Txo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "outpoint",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "script",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/api/txos/outpoints": {
			"post": {
				"operationId": "PostOutpoints",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Txo"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "script",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/api/inscriptions/{outpoint}": {
			"get": {
				"operationId": "GetTxoByOutpoint",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Txo"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "outpoint",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "script",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			}
		},
...
		"/api/bsv20/{address}/balance": {
			"get": {
				"operationId": "GetBalanceByAddress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TokenBalanceResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "address",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}